Cleanup, move some code to a g file, fix mushroom btn bug, reduce repetitiveness and improve perf with switch statements
