+++ src/global/commands/g.search.ts
@@ -1,0 +1,275 @@
+import { stripIndents } from 'common-tags';
+
+export async function normalDefine(term: string): Promise<{
+  title: string, url: string, description: string
+}> {
+  try {
+    const url = `https://api.dictionaryapi.dev/api/v2/entries/en/${encodeURIComponent(term)}`;
+    const response = await fetch(url);
+    if (!response.ok) {
+      return {
+        title: `Dictionary: ${term}`,
+        url: `https://www.google.com/search?q=define+${encodeURIComponent(term)}`,
+        description: `No dictionary definition found for "${term}".`,
+      };
+    }
+    const data = await response.json();
+    if (!Array.isArray(data) || !data[0]?.meanings?.length) {
+      return {
+        title: `Dictionary: ${term}`,
+        url: `https://www.google.com/search?q=define+${encodeURIComponent(term)}`,
+        description: `No dictionary definition found for "${term}".`,
+      };
+    }
+    const meaning = data[0].meanings[0];
+    const { definition } = meaning.definitions[0];
+    const example = meaning.definitions[0].example
+      ? `\nExample: ${meaning.definitions[0].example}`
+      : '';
+    const partOfSpeech = meaning.partOfSpeech
+      ? `*(${meaning.partOfSpeech})* `
+      : '';
+    return {
+      title: `Dictionary: ${term}`,
+      url: `https://www.google.com/search?q=define+${encodeURIComponent(term)}`,
+      description: stripIndents`
+        ${partOfSpeech}${definition}${example}
+      `,
+    };
+  } catch (err) {
+    return {
+      title: `Dictionary: ${term}`,
+      url: `https://www.google.com/search?q=define+${encodeURIComponent(term)}`,
+      description: 'Dictionary API is currently unavailable. Please try again later.',
+    };
+  }
+}
+
+export async function urbanDefine(term: string): Promise<{
+  title: string, url: string, description: string
+}> {
+  try {
+    const url = `https://api.urbandictionary.com/v0/define?term=${encodeURIComponent(term)}`;
+    const response = await fetch(url);
+    if (!response.ok) {
+      return {
+        title: `Urban Dictionary: ${term}`,
+        url: `https://www.urbandictionary.com/define.php?term=${encodeURIComponent(term)}`,
+        description: 'Urban Dictionary is currently unavailable. Please try again later.',
+      };
+    }
+    const data = await response.json();
+    if (!data.list || data.list.length === 0) {
+      return {
+        title: `Urban Dictionary: ${term}`,
+        url: `https://www.urbandictionary.com/define.php?term=${encodeURIComponent(term)}`,
+        description: `No results found for **${term}**.`,
+      };
+    }
+    const entry = data.list[0];
+    return {
+      title: `Urban Dictionary: ${entry.word}`,
+      url: entry.permalink,
+      description: stripIndents`
+        ${entry.definition.replace(/\[([^\]]+)\]/g, '$1')}
+
+        *Example:* ${entry.example.replace(/\[([^\]]+)\]/g, '$1')}
+        👍 ${entry.thumbs_up} | 👎 ${entry.thumbs_down}
+      `,
+    };
+  } catch (err) {
+    return {
+      title: `Urban Dictionary: ${term}`,
+      url: `https://www.urbandictionary.com/define.php?term=${encodeURIComponent(term)}`,
+      description: 'An error occurred while searching Urban Dictionary.',
+    };
+  }
+}
+
+export async function steam(query: string): Promise<{
+  title: string,
+  url: string,
+  description: string,
+  thumb: string,
+  header?: string,
+  fields: { name: string, value: string, inline?: boolean }[]
+}> {
+  const url = `https://store.steampowered.com/api/storesearch/?term=${encodeURIComponent(query)}&cc=us`;
+  const response = await fetch(url);
+  if (!response.ok) {
+    return {
+      title: `Steam: ${query}`,
+      url: `https://store.steampowered.com/search/?term=${encodeURIComponent(query)}`,
+      description: `No Steam results found for "${query}".`,
+      thumb: '',
+      fields: [],
+    };
+  }
+  const data = await response.json();
+  if (!data.items || data.items.length === 0) {
+    return {
+      title: `Steam: ${query}`,
+      url: `https://store.steampowered.com/search/?term=${encodeURIComponent(query)}`,
+      description: `No Steam results found for "${query}".`,
+      thumb: '',
+      fields: [],
+    };
+  }
+  const item = data.items[0];
+
+  const currencies = [
+    { cc: 'us', symbol: '$' },
+    { cc: 'de', symbol: '€' },
+    { cc: 'gb', symbol: '£' },
+  ];
+  const details: Record<string, any> = {};
+  const detailsResults = await Promise.all(
+    currencies.map(cur => fetch(`https://store.steampowered.com/api/appdetails?appids=${item.id}&cc=${cur.cc}`)
+      .then(res => res.json())
+      .then(json => ({ symbol: cur.symbol, data: json[item.id]?.data }))
+      .catch(() => ({ symbol: cur.symbol, data: undefined }))),
+  );
+  detailsResults.forEach(({ symbol, data: currencyData }) => {
+    details[symbol] = currencyData;
+  });
+
+  const usDetails = details.$ || {};
+  type Genre = { id: number; description: string };
+  let tags: string;
+  if (usDetails.genres && Array.isArray(usDetails.genres)) {
+    tags = (usDetails.genres as Genre[]).map(g => g.description).join(', ');
+  } else if (item.tags && Array.isArray(item.tags)) {
+    tags = item.tags.join(', ');
+  } else {
+    tags = 'N/A';
+  }
+  const header = usDetails.header_image || item.header_image || null;
+
+  let releaseDate = 'Unknown';
+  if (usDetails.release_date?.date) {
+    if (usDetails.release_date?.coming_soon === false && usDetails.release_date?.steam_release_date) {
+      releaseDate = `<t:${usDetails.release_date.steam_release_date}:D>`;
+    } else {
+      releaseDate = usDetails.release_date.date;
+    }
+  } else if (item.release_date) {
+    releaseDate = item.release_date;
+  }
+
+  const description = usDetails.short_description || item.name;
+  const developer = usDetails.developers ? usDetails.developers.join(', ') : 'Unknown';
+  const publisher = usDetails.publishers ? usDetails.publishers.join(', ') : 'Unknown';
+
+  let priceFieldName = 'Price';
+  let salePercent = 0;
+  const usdPriceObj = details.$?.price_overview;
+  if (usdPriceObj && usdPriceObj.discount_percent && usdPriceObj.discount_percent > 0) {
+    salePercent = usdPriceObj.discount_percent;
+    priceFieldName = `Price (${salePercent}% off)`;
+  }
+  const prices = currencies.map(cur => {
+    const priceObj = details[cur.symbol]?.price_overview;
+    return priceObj ? `${cur.symbol}${priceObj.final_formatted.replace(/[^0-9.,]/g, '')}` : `${cur.symbol}N/A`;
+  }).join(', ');
+
+  return {
+    title: `Steam: ${item.name}`,
+    url: `https://store.steampowered.com/app/${item.id}`,
+    description,
+    thumb: item.tiny_image,
+    header,
+    fields: [
+      { name: 'Tags', value: tags, inline: true },
+      { name: 'Release', value: releaseDate, inline: true },
+      { name: '\u200B', value: '\u200B', inline: true },
+      { name: 'Developer', value: developer, inline: true },
+      { name: 'Publisher', value: publisher, inline: true },
+      { name: '\u200B', value: '\u200B', inline: true },
+      { name: priceFieldName, value: prices, inline: false },
+    ],
+  };
+}
+
+export async function wikipedia(query: string): Promise<{
+  title: string, url: string, description: string, thumb: string
+}> {
+  const url = `https://en.wikipedia.org/api/rest_v1/page/summary/${encodeURIComponent(query)}`;
+  const response = await fetch(url);
+  if (!response.ok) {
+    return {
+      title: `Wikipedia: ${query}`,
+      url: `https://en.wikipedia.org/wiki/${encodeURIComponent(query)}`,
+      description: 'Wikipedia is currently unavailable. Please try again later.',
+      thumb: '',
+    };
+  }
+  const data = await response.json();
+  return {
+    title: `Wikipedia: ${data.title || query}`,
+    url: data.content_urls?.desktop?.page || `https://en.wikipedia.org/wiki/${encodeURIComponent(query)}`,
+    description: data.extract || '',
+    thumb: data.thumbnail?.source || '',
+  };
+}
+
+export async function weather(city: string): Promise<{
+  title: string,
+  url: string,
+  description: string,
+  fields: { name: string, value: string, inline?: boolean }[]
+}> {
+  try {
+    const geoRes = await fetch(`https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(city)}`);
+    const geoData = await geoRes.json();
+    if (!geoData[0]) {
+      return {
+        title: `Weather: ${city}`,
+        url: `https://wttr.in/${encodeURIComponent(city)}`,
+        description: 'Could not find this location. Try being more specific',
+        fields: [],
+      };
+    }
+    const { lat, lon, display_name: displayName } = geoData[0];
+
+    const wttrUrl = `https://wttr.in/${lat},${lon}?format=j1`;
+    const weatherRes = await fetch(wttrUrl);
+    const weatherData = await weatherRes.json();
+
+    const current = weatherData.current_condition?.[0];
+    const weatherDesc = current?.weatherDesc?.[0]?.value || 'N/A';
+    const tempC = current?.temp_C ?? 'N/A';
+    const tempF = current?.temp_F ?? 'N/A';
+    const feelsLikeC = current?.FeelsLikeC ?? 'N/A';
+    const feelsLikeF = current?.FeelsLikeF ?? 'N/A';
+    const windKph = current?.windspeedKmph ?? 'N/A';
+    const windDir = current?.winddir16Point ?? '';
+    const humidity = current?.humidity ?? 'N/A';
+    const precipMM = current?.precipMM ?? 'N/A';
+
+    const mapsUrl = `https://www.google.com/maps/search/?api=1&query=${lat},${lon}`;
+    const description = `[View on Google Maps](${mapsUrl})`;
+
+    const fields = [
+      { name: 'Condition', value: weatherDesc, inline: true },
+      { name: 'Temperature', value: `${tempC}°C / ${tempF}°F`, inline: true },
+      { name: 'Feels Like', value: `${feelsLikeC}°C / ${feelsLikeF}°F`, inline: true },
+      { name: 'Wind', value: `${windKph} km/h ${windDir}`, inline: true },
+      { name: 'Humidity', value: `${humidity}%`, inline: true },
+      { name: 'Precipitation', value: `${precipMM} mm`, inline: true },
+    ];
+
+    return {
+      title: `Weather: ${displayName}`,
+      url: wttrUrl,
+      description,
+      fields,
+    };
+  } catch (err) {
+    return {
+      title: `Weather: ${city}`,
+      url: `https://wttr.in/${encodeURIComponent(city)}`,
+      description: 'Weather is currently unavailable. Please try again later.',
+      fields: [],
+    };
+  }
+}
